Share params between net/http and gin
